[MSBuild] Add basic MSBuild tasks for generating a NuGet package
This adds minimal support for generating a NuGet package using
metadata in a project. It is based on the existing tasks from NuProj.
It does not automatically generate all the information required in
a NuGet package from the project or any referenced projects. It adds
two new MSBuild targets GetNuGetInfo and BuildNuGet with a very basic
implementation.