[MSBuild] Do not build NuGet package if no metadata available
If the NuGet package id is not set in the project's properties then
the NuGet package will not be built when BuildNuGet is set to true.
This allows the entire solution to be built with BuildNuGet set to
true and only those projects that are configured to generate NuGet
packages will do so. Previously all projects would attempt to
generate a NuGet package when BuildNuGet was set to true.