[MSBuild] Handle project references when generating a NuGet package
For referenced projects the GetNuGetInfo MSBuild task is called.
This generates a .nuspec file for each project reference. The .nuspec
files are passed to the GenerateNuSpec MSBuild task so it can
directly include the referenced project assemblies into its own
NuGet package, or remove the referenced project assemblies when the
referenced project is configured to create its own NuGet package.