[MSBuild] Handle ReferenceOutputAssembly false for project references
When ReferenceOutputAssembly is false for a project reference the
output assembly is not referenced from the referenced project. When
building a NuGet package for a project that has a reference with
ReferenceOutputAssembly is false the project reference is ignored.
Previously any NuGet package dependencies would be added to the
generated NuGet package for these project references.