[MSBuild] Fix incremental build
The .nuspec file was being generated on every build even if nothing
in the .nuspec file would be changed. This was because the
GenerateNuSpec MSBuild task was comparing the original .nuspec file
contents with an in memory stream that still had the byte order mark
at the start of the string. Using a StreamReader removes this byte
order mark so the string comparison now works.

Also added tests incremental builds.