[MSBuild] Support .nuproj project referencing another .nuproj
When building a .nuproj project that references another .nuproj
project the referenced project's NuGet package will be added as a
dependency in the referencing project. The NuGet package dependency
is not associated with any target framework. One use for this would
be to create a metapackage which has no content itself but just
provides a way to install other NuGet packages.