[MSBuild] Fix test failures on Windows
No longer setting BuildingInsideVisualStudio to true when building the
solution. On Windows when this is set to true the projects were not
built in the correct order but seemed to be built in the order defined
in the solution so referenced project may not be built when a project
is built causing the build to fail.

Fixed GetNuGetInfo not returning any information for project
references by adding a Returns parameter and returning a new NuGetInfo
item with the correct path to the .nuspec file.

Fixed .nuproj builds failing since the Microsoft.Common.targets tries
to copy the generated .nupkg from obj\Debug before the NuGet package
is created, although it is not created in that directory but directly
built into the bin\Debug directory. This was fixed by overriding
PrepareForRunDependsOn so CopyFilesToOutputDirectory is not run.

Parallel xunit test runs are now disabled. Some of the test runs would
fail since the directory containing the test projects being built was
being deleted whilst another test was running. This never happened on
the Mac but happens on Windows.