[MSBuild] Mark required properties for GenerateNuSpec task
All required properties for the GenerateNuSpec task are now marked
with the [Required] attribute. The required information is: id,
version, description and author. Without this information it is
possible to save the .nuspec file but then it cannot be read again
by NuGet since it validates the information on reading. Previously
the GenerateNuSpec task added place holder values for the required
metadata for projects that did not provide it. Now this place holder
values are set in the common MSBuild targets instead.