[MSBuild] Use MSBuild instead of xbuild when running tests
Switch to using MSBuild instead of xbuild since MSBuild is now
included with Mono.

The GenerateNuSpec task has to change to handle Mono's MSBuild not
converting metadata values so they contain the correct directory
separators for the current platform. XBuild would convert the
metadata values and fix the directory separators.