Don't bring nuget references from resolved assembly references
Even if it is convenient and easy, it isn't what we need. Resolved
nuget assets include the actual references from all transitive
dependencies of the direct dependencies a project takes. For example,
in the skipped test we added, the csproj has a project.json with a
dependency on Microsoft.CodeAnalysis, but the resulting project.lock.json
ends up having all the transitive dependencies of that package, and
not a single trace that the original dependency was on Microsoft.CodeAnalysis,
since it is a meta-package :).

So this approach isn't useful for nugetizer. We'll need to properly
read in the project.json and packages.config eventually.