Removed duplication of manifest metadata, complete e2e with references
CreatePackage now is way simpler: we just pass in the manifest item we
create automatically for it, so there is no need to duplicate properties.

Output package is returned as an item with all the original metadata
used from its manifest too, for easier tooling too.

Added files to generated .nupkg from current package and transitive
references.

Renamed BuildPackage and Pack project-level properties to PackOnBuild,
which is more explicit on what it does.

An entire solution's packages can be built by doing:

    msbuild my.sln /p:PackOnBuild=true

Added sample solution to a realistic graph of project dependencies to
Scenarios/given_a_complex_pack/pack.sln, which is exercised in tests
but can also be directly built from the source.

Partially implements NuGet/Home#3477