Bootstrap NuGet.Build.Packaging nuget itself
The update to the build.proj causes the Package target
to build the solution passing /p:PackOnBuild=true, which
causes the Tasks' project custom targets to provide the
PackageFiles we want to include in the final nuget.

Shared package metadata (typical in most projects, things
like owners and authors, copyright and urls) is provided
via a Shared.targets, to also showcase a common pattern.