Reorganize targets to better align with spec
GetPackageContents now returns the same filtered content that would
end up in the .nupkg. The new GetAllPackageContents provides the
unfiltered transitive content for the current project and all references.

Also made it possible to always get the package path and version regardless
of whether it was set already or not (separating the _SetDefault* in a
separate target).