Simplify build/restore for nuget package for the VSIX
We previously had a complicated dependency between the VSIX and the
build tasks so that we could build both as a project reference during
command line builds, but have a project reference at design-time.

There wasn't in the end much value in having the Tasks project in the
VSIX solution, in fact, the contrary, since the tasks is backwards
compatible with VS2015, whereas the VSIX targets VS2017 exclusively
and therefore has incompatible MSBuild nuget package dependencies.

By removing the project reference and simplifying the command line
build, we just require the build to exist (building once the Build
solution or build.cmd), we remove the complexity and make it easier
to maintain moving forward.

The restore has also switched to using MSBuild built-in restore
capability, instead of using nuget.exe for projects.