Use $(IsPackable) everywhere we read $(PackageId)
At the top of the targets file, we default `$(IsPackable)` to true if `$(PackageId)` is
not empty. .NETStandard projects currently *always* assign a value to that property,
matching the `$(AssemblyName)`, therefore breaking our logic.

To improve compatibility, we can check for the `$(IsPackable)` property instead, so that
while more alignment is achieved, at least an explicit `$(IsPackable)==false` can be set
for those projects and keep our logic consistent in that case.

That involves never assuming the `$(PackageId)` will be empty before assigning it as
metadata on items, and always checking for `$(IsPackable)==true` instead.