Allow overriding authoring package TF/restore behavior from nugetizer
Even if we do need to provide these properties in the unfolded .nuproj
anyway so that Restore works on CI, we can still provide better values
over time after the initial restore.

See also https://github.com/NuGet/Home/issues/5254 for a better way in
the future to handle this.