Improvements to Content management for consistency
Prevent users from using `contentFiles` as a relative dir in their
@Content items, since that would be a user error since the folder
is already added to Content items with no CopyToOutputDirectory.

Keep the original behavior where Content is by default:
- CodeLanguage=any
- TargetFramework=current TFM.

Meaning that Content is by default %(FrameworkSpecific)=true. This
can obviously be overriden by either:

- Adding an explicit %(TargetFramework) to target (i.e. 'net46' or 'any')
- Adding %(FrameworkSpecific)=false

Turned off inclusion of None items by default, since they basically serve
the same purpose as raw PackageFile. Turning it on by default could mean
inadvertently including stuff that we shouldn't.

Included additional tests for the metadata overrides for lang/tfm.

NOTE: we're not defaulting %(Content.TargetFramework) via an
ItemDefinitionGroup to avoid potentially colliding with other targets in
what seems like a super generic metadata name. So instead, we process
the default in AssignPackagePath and assign that default value to the
resulting items, so the calculated/inferred value can be inspected if
necessary.

Also, added the PackageFile as an AvailableItemName so that users can
trivially select it as the Build Action for an item, and cause it to
be included in the package by default. This would make using None
completely unnecessary.