Properly add Clide as a dependency (#139)
When we bumped VsixDependency to properly account for VS2017
pre-requisite, we lost the plain Dependency nodes for non-component
VSIXes (Clide in our case). This caused our MEF exports to not be
available and the NuGet CPS exports to be missing, causing the nuproj
to fall back to packages.config behavior, which doesn't work.