Add the prerequisite Merq component explicitly
The existing implicit prerequisite insertion from Merq nuget package isn't working
as expected because the component ID in VS doesn't match the generated JSON file name,
so instead of relying on the Merq.VisualStudio nuget package to inject the prerequisite,
we just add it ourselves manually.

Also fixed versioning of the nuget package for local builds with the new numbering scheme.

Fixes #5383