Added support for opening binlog on when test build runs
A new `[OpenBuildLog]` can be used to annotate `[Fact]/[Theory]` tests so that
right after a build happens, the automatically generated .binlog is opened using
Process.Start. This will only work if the viewer is installed from http://msbuildlog.com.

The log is not generated unless the attribute is present, so there's no overhead
unless it's used. Also, even if you forget the attribute and commit/push, on
release builds, the attribute is ignored and the log is not launched.