Fix failing test due to nuget package path casing
The package cache lower cases all paths, which no longer matches the
package id as we were previously doing. We need to match in a case-
insensitive way therefore for private assets matching to succeed.

This fixes a broken unit test that was blocking two PRs #151 and #150.

Also, fix the nuget package versions at the time of this test authoring,
so we don't accidentally break if newer packages ever change and cause
the expectations to fail.