Leverage module initializer to set MSBuild path
Figured out that the MSBuild assemblies were being loaded too early because of the LoggerVerbosity attribute we had to make it easier to bump logging when troubleshooting a test. By making the attribute internal, the MSBuild assemblies aren't now loaded anymore by the test runner, which now allows us to use the proper MSBuild locator to set the path to the one used to build the project.

A cleanup step is needed because MSBuild logger and the tasks themselves cause the MSBuild assemblies to be copied to the output directory, so we need to remove them so they don't get accidentally loaded from there.