Explicit PackagePath in Content/None equals Pack=true
This makes it more intuitive and aligns with SDK Pack.
Also, since de-duplication is performed in the CreatePackage task, the targets were being overzealous with that and preventing the valid scenario (like we have ourselves!) of using the same source file and place it in different package paths (we do so for various multi-targeting ones.