Added tests for content files metadata in nuspec
Content files can have additional metadata that controls how they are included in the referencing project. See https://docs.microsoft.com/en-us/nuget/reference/nuspec#using-the-contentfiles-element-for-content-files.

These tests ensure we set the right metadata in the nuspec we generate.

Added the nuspec full path as metadata to the output item from CreatePackage if nuspec generation is opted-in, which now allows us to test the final nuspec without generating the full nupkg too :) (leveraging something the dotnet-nugetize tool already uses for the same purpose).