Add support for packing sources
The behavior is similar to Content: if the packed items have CopyToOutputDirectory, they become part of the build output kind. Otherwise, they are treated like content with a default CodeLanguage that matches the project's default language source extension. This makes the most sense and would in most cases not require any futher metadata annotations.

Tests ensure that the overriding of CodeLanguage, TargetFramework and other metadata still work as expected.