Add remaining pack inference items from SDK pack
The test showcases how the mechanism is itself extensible (which is what we basically did with the new built-in ones) and how the same rules can be applied to any custom item types.

Corresponding PackXXX properties were added for each new item type, so they can also be defaulted to Pack=true (or false) automatically like the existing Content/Compile/None/EmbeddedResource.