Rename Kind to PackFolder
Kind as metadata for an item does not look very idiomatic. Since every other property and metadata starts with `Pack`, this renames it to `PackFolder`. The `PackFolderKind` items define the defaults now.

The compatiblity targets were extended to account for the now-legacy values from previous releases as well as the pre-fork nugetizer.

Update the docs accordingly too.