Add support for PrivateAssets=all on ProjectReference too
This is a must-have for build and analyzer packages which need to pack their dependencies privately, yet they are authored typically in the same solution as the main library they contribute to, and might need its assets as private (i.e. in an analyzer that uses the types in the main library).