Add support for centrally managed package versions
See feature spec on NuGet at https://github.com/NuGet/Home/wiki/Centrally-managing-NuGet-package-versions.

This feature is shipping (with partial IDE and CLI support, but fully working in MSBuild already) in current .NET5 previews, and it's stable and useful enough to support in NuGetizer built-in.

These changes make it possible to automatically pack (as dependency or private assets) PackageReference which don't specify a Version because it's done in a Directory.Packages.props file instead.