Add support for pass-through MSBuild properties
Leverage that same mechanism for our own built-in properties. This also means we don't need to explicitly run restore, since users can just pass `-r` and we'll pass-through that to MSBuild to do its thing.

Also make the syntax for binlog arg the same as MSBuild (`-bl`) so that we're effectively a superset of that switch (we also automatically attempt to open the log file after the run, if msbuildlog.com is installed).

Also, the --version switch will cause us to also render MSBuild logo/version info (without it, we pass -nologo instead), unifying the experience there too.