By default, packing Compile should not include generated files
Like the target framework attribute and assembly info.
Additional exclusions should be configurable as needed, so we moved the <PackInference> item to a .props file so the project can update/remove as needed.

This adds another customization point too, allowing to tweak the exclusions for any PackInference (i.e. for Content, EmbeddedResource, None and so on).