Allow rendering contributed contents from nugetize
If the project was not packable, we were not rendering anything, which wasn't
very useful when tweaking a project that contributes to a parent package.

This adds that capability by explicitly re-running the GetPackageContents target
with additional indication that we're collecting contents, so that the package path
assignment happens forcedly. Note that this does not account for the potential
retargeting of contents if the referencing project happens to redefine the TF
(i.e. contributing a NS2 library, but referencing project is net472+net5.0, the
lib will actually contribute to both TFMs).

Also make it possible to select a specific project if the directory contains
more than one, by smartly parsing the extra arguments.