When referencing relative package content, ensure full path from referencing projects
Given a packaging project, the PackageFile items determined (either explicitly or via inference) will almost always be relative paths. A packaging project will therefore fail to pack in those cases, unless we turn those files into full paths.

For that purpose, we append a new metadata always to all _PackageContent determined in the GetPackageContents target: MSBuildSourceProjectFile (follows same name as when Rebase=true is used in an MSBuild task execution) and the additional MSBuildSourceProjectDirectory. The former provides consistency, the latter can be used to create the full original item path by combining it with OriginalItemSpec. The actual work to make the item full path is done in the AssignPackagePath task, which is invoked always before returning from GetPackageContents.

An escape-hatch mechanism is added nevertheless, where you can annotate the content with `%(IsFile)=false` to avoid it being treated as one.