Update and future-proof build scripts
First and most important: add unicode icons!
Second: instead of hardcoding `preview` as the VS being searched, leverage the ability to select the first instance greater than or equal to 16.8 we find on the machine. This means we won't be force-installing a preview VS when 16.8 becomes the built-in version in hosted agents. Also, it means when 16.9 or future versions become stable too, we won't need to modify scripts again, since we'll automatically use a later version this way.