When inferring private dependencies, preserve item metadata
The item metadata on a `PackageReference` may be useful to consumers that use `PrivateAssets` to automatically add implicit package references to transitive dependencies (i.e. to include all dependent assemblies as part of a build tasks assembli). The user-defined metadata may later be used to further tweak the inference (i.e. selectively removing some items from the resulting `PackageFile` items). For this scenario, it's useful to preserve the original metadata in the top-most `PackageReference` that triggers an implicit transitive dependency tree to be a candidate for inference.

We also remove from the task the conditional checks on the package reference metadata since the targets already performed the same checks. This makes the task more flexible, should a user decide to invoke it explicitly as part of some custom inference they might want to perform (i.e. inferring content based on different metadata than `PrivateAssets=all`).