Fix issue with usage of PackOnBuild on multitargeting projects
Due to the way we were setting up the BuildDependsOn when PackOnBuild == true, we were running Pack after build, regardless of whether the build was single or multi-targeting. This caused the Pack to run once for each TF, causing each build to overwrite the previous package and therefore leaving only the last TFM built as the package contents.

This change switches to using Before/After targets and conditions instead with DependsOn for Build/Pack, as well as a mirror property to PackOnBuild, BuildOnPack to avoid building before pack (which the default behavior in SDK pack too).

Fixes #32.