Convert ref to lib when packing transitive private dependencies
When packing a package reference with PrivateAssets=all, you typically
do so in order to run the resulting code from the target pack folder
directly without having any extra dependencies that have to be resolved
from elsewhere. For this typical use case, you'll want to pack actual
"runnable" assemblies, not reference assemblies.

Packages that pack reference assemblies (typically done for netstandard)
resulted code that couldn't run because the packed assemblies contained
no code.

So we first look at related files that have a "/ref/" in the path, see
if we can find the same file but under "/lib/" and if so, pack that
instead.

Fixes #36