Import compatibility props for multi-targeting too
Because we were importing the compatibility props only from the non-multitargeting targets, the defaulting of PackOnBuild from GeneratePackageOnBuild was not being applied and that caused usage of that property to fail in multi-targeting projects.

By bringing in the compat props from the shared targets (used in both single and multi targeting scenarios), we can properly default all values in all types of builds.

Fixes #34