Normalize all paths for consistency, use SDK Pack format
We used to perform no path normalization in the AssignPackagePath task, and this meant we could easily end with mixed paths, making it difficult to tell apart which are the same, and when comparing via code, having to take into account the possibility that there might be a mixture.

To simplify things, we now normalize all paths to use the / forward slash which is more compatible for cross-platform projects, is already supported universally in MSBuild as well as NuGet, and entirely avoids having to do \\ double slashes for escaping (or @ literals) in code.

Fixes #135