Delete generated packages explicitly when clean
Generated NuGet packages will be recorded in "FileWrites" items in
"NuGetizer.Shared.targets" so they can be deleted when clean.
However, that method only works for files in the intermediate
directory, which can easily happen if user set "PackageOutputPath"
outside of that directory.

To fix, call "CleanPackageOutput" target to delete those files
explicitly when clean.