Bump tests to net6 too, restore one NoTargets before tests
More net6 updates on test projects. We also had local failures when the NoTargets MSBuild SDK wasn't found, so we do an explicit restore after build and before tests, so it's properly located when loading projects.

Revert test project to net472 since that's what was last working on CI/main.