Build with MSBuild so the proper assemblies are referenced
When running tests with dotnet test, the x86 test host will run, which will run on the desktop CLR, not the dotnet core CLR, and therefore we need the build to also point to the desktop assemblies which live in the VS installation used by msbuild. We can force all that by using msbuild instead of dotnet build for building, since we're now using windows-only build anyway too. (NOTE: this might still work on linux/macOS if the setup msbuild task does The Right Right).