Be smarter about PackFolder on project references
Whenever we had `PrivateAssets=all` on a project reference (whether implicitly or explicitly), we used to unconditionally pass `PackFolder=$(PackFolder)` when resolving referenced content. This is correct whenever the calling project does indeed have a `PackFolder`, but not for packaging projects, since in that case, the packaging project itself doesn't have a value for that property (since it can pack any arbitrary content and has no primary output).

So we switch to the mechanism used already by the .NET SDK (and common targets) of using a metadata item with the full property assignment syntax. This means that if the metadata isn't present (which we now condition to the presence of a value for `PackFolder`), when calling `GetPackageContents` on the project reference, no value will be passed at all for that property, meaning whichever value is there already, will be propagated to the referencing project, which is exactly what we want in a packaging project scenario.

Fixes #139