Allow packaging projects to be framework-specific
We refine the support for packaging projects in this commit, by defaulting its PackFolder to Ignore, which becomes somewhat similar to None, but allows for differenciation.

Moreover, we test specific scenarios involving a packaging project becoming framework-specific (which we previously didn't explicitly support) by means of setting `BuildOutputFrameworkSpecific=true` at the project level. This requires re-targeting referenced assets whenever that property is true, in addition to non-packaging projects in that same scenario.

Fixes #182