Improve default heuristics for build projects
Make it so that the default behavior for None items when the PackFolder for the project is build or buildTransitive the None matches the project. This is unless the item doesn't specify a different FrameworkSpecific or PackFolder metadata.

This fixes the default items that were not being properly populated for packaging projects too, due to the lack of a default file extension (meaning all default items end up being ignored).

Fixes #184.