Improve readme content inclusion metadata inference
A readme.md file at the project level will typically be automatically included as a None item,
which by default won't pack either.

This improves the inference by adding a `PackReadme=true` (by default) which controls whether a readme file (either explicitly specified with `PackageReadmeFile` as per the SDK Pack spec, or implicitly found alongside the project) is automatically packed and set as the Readme package metadata (as long as the project is being packed (IsPackable=true)).

NOTE: this is implemented in the Inference targets, so that when inference is turned off entirely,
this doesn't happen automatically, for consistency with other items inferences.

Fixes #185