Improve compatibility with SDK Pack for tools
This brings nugetizer pretty much to parity with SDK Pack for CLI tools too :).

We leverage the SDK tool packing targets to determine the files to pack, but otherwise preserve nugetizer's inference and packing behavior. Best of both worlds!

The new `InferToolContents` target is the one that transforms the `PackTool` target's `TfmSpecificPackageFile` into our own `None` items for packing. It can also be extended to trim the tool's contents via a custom target that removes items, such as::

```xml
  <Target Name="TrimRuntimes" AfterTargets="InferToolContents">
    <ItemGroup>
      <None Remove="@(None)" Condition="$([MSBuild]::ValueOrDefault('%(None.PackagePath)', '').Contains('/runtimes/'))" />
    </ItemGroup>
  </Target>
```

(this target would remove all runtimes from the tool package, but otherwise preserve the package contents).

Fixes #134