Support transitive package dependency from P2P
We were previously not re-targeting dependencies from referenced (non-packing) projects, which would end up as dependencies in the originating project TFM, instead of the packing (TFM-specific) project, causing a runtime failure since the dependency wouldn't be present, even if the referenced assembly would.

This is now supported, including in multi-targeting scenarios for the packing project.

Fixes #199