Automatically resolve includes in package readme file
Allows a package readme to simply reference other files (via relative path and optional
fragment specifier), like:

```
<!-- include ../../readme.md#content -->
<!-- include ../../docs/footer.md -->
```

Where the main project readme could contain a bunch of project-level info,
contrib, badges, etc., which you can skip by marking a specific area as t
he #content anchor:

```
# My Project
// bunch of badges
// building/contributing, etc.

<~-- #content -->
# Usage
// now the real stuff we want in a package readme
```

The `footer.md` could contain project sponsors and the like, and also be
reused across multiple packages.

The includes must be relative (to the including file) file paths, and the fragmnet
specifier is optional. If the fragment specifier does not have an "end" anchor in
the included file, the file is included from the anchor definition until EOF.

The includes are only resolved when the actual .nupkg is created, and via a temp
file, to avoid affecting the build's incrementality (which doesn't apply to the .nupkg
itself either).

Fixes #210