Don't fail nugetize for tools
For dotnet tool projects, the design-time build we performed failed since
the tool attempted to publish its output, which the project references weren't
able to supply and the whole thing failed with missing files.

We now pass IsPublishable=false to fake a sort of "DTB" for tools too,
so they don't attempt to publish output.

Fixes #247