Add NuGetPackageId metadata to target path item
This allows P2P references to inspect this value in a standard way that is consistent with package references.

Resolved assets via NuGet always contain this item metadata, so it makes sense to add it to a referenced project's output for consistency. In the case of analyzers, this could be helpful in filering analyzers by their package id in both cases in a uniform way.