Pack runtime lib assets when inferring dependencies
We were previously using a combination of @(ReferencePath) and @(_ReferenceRelatedPaths) to determine the candidate libs to pack as private, but this was incorrect since we would still sometimes pack ref assemblies.

Turns out that this is an issue that surfaced elsewhere (see https://github.com/NuGet/Home/issues/9310#issuecomment-612307251) which resulted in a new output group being available from the common MSBuild targets (see https://github.com/dotnet/msbuild/issues/3069) that we can use instead to do this properly.

Closes #263