Fix support for platform-suffixed build output
Turns out that the new-ish platform-suffixed TF names (i.e. net7.0-windows, net7.0-maccatalyst and so on) all end up being the same TFM (.NETCoreApp, Version=7.0), which ends up packing assets to duplicate package paths in pack inference.

It seems like we might need to rethink nugetier reliance on the TFM for most of its multi-targeting and TF-specific packing. We currently use the TFM as the default to determine target package path, unless a TF is specified for an item. This ends up being incorrect if the TF contains a platform suffix, as noted.

For now, provide a specific and localized fix for just primary output assembly, symbols and docs.

We should do an extensive examination of other scenarios for xplat MAUI packages.