Change nugetize default to non-verbose
We were showing MSBuild output for any nugetize run that took longer than 10 seconds. Diagnosing slow builds shouldn't be a nugetizer issue, and it makes it confusing to see potentially many warnings and what-not that are irrelevant to packing.

Make it so instead of asking for quiet, you can ask for verbose and get the output.

If there were errors running the tool, we'd still want to see the MSBuild output, though.