Use a better CI version number
By making the last number the build number, we ensure we have ever-incrementing numbers, even across branches. This makes it easier to have wildcard versions too, and the prerelease label becomes just the branch name.

This should make the targets that use VersionLabel entirely unnecessary, so we might want to make that change in the oss template as a single commit, so after the bump, both the simplified targets and new version # kicks in.