Add compatibility with SDK package validation
Package validation is increasingly important for multi-targeting libraries, so NuGetizer must be compatible with this feature which became first available in the .NET6 SDK.

In addition to adding the basic compatibility, we also turn the feature on by default (setting EnablePackageValidation=true) for multi-targeting projects when building in Release mode, to ensure best-practices are applied by default.

Users can opt-out of this by setting it to `false` instead explicitly.